third try to fix this stuff

+++ src/js/attribute_widget/SelectionValue.js
@@ -89,8 +89,9 @@ define([
                     _iwc.setBufferedMessagesReceiver(CONFIG.WIDGET.NAME.VIEWCANVAS);
                     var viewTypeAttribute = viewType.getAttributes()["[attributes]"];
                     var viewTypeAttrList = viewTypeAttribute.getAttributes();
+                    var attributeList = null;
 
-                    if(!fromCallback) {
+                    if (!fromCallback) {
                         //delete the old all attributes of the old target
                         for (var key1 in viewTypeAttrList) {
                             if (viewTypeAttrList.hasOwnProperty(key1)) {
@@ -99,7 +100,9 @@ define([
                                 attr.propagateAttributeDeleteOperation(op, CONFIG.WIDGET.NAME.VIEWCANVAS);
                             }
                         }
-                        var attributeList = {};
+
+
+                        attributeList = {};
                         //the attributes of the new target
                         var targetAttributes = node.getAttributes()["[attributes]"].getAttributes();
                         //crate the attributes of the new target
@@ -120,6 +123,15 @@ define([
                             }
                         }
                     }
+                    //if from callback is true, attributeList is null. Initialize it
+                    if(!attributeList){
+                        attributeList = {};
+                        for(var key in viewTypeAttribute){
+                            if(viewTypeAttribute.hasOwnProperty(key)){
+                                attributeList[key] = viewTypeAttribute[key].getKey().getValue();
+                            }
+                        }
+                    }
                     if (condListAttr = viewType.getAttribute('[condition]')) {
                         condListAttr.setOptions(attributeList);
                         var attrList = condListAttr.getAttributes();
@@ -131,20 +143,20 @@ define([
                             }
                         }
                     } else {
-
                         var condAttr = new ConditionListAttribute("[condition]", "Conditions", viewType, attributeList, LogicalOperator, LogicalConjunctions);
                         viewType.addAttribute(condAttr);
                         viewType.get$node().find('.attributes').append(condAttr.get$node());
                     }
 
-                    EntityManager.deleteFromMap(viewType.getViewId(), viewType.getEntityId());
-                    EntityManager.addToMap(viewType.getViewId(), node.getEntityId(), viewType.getEntityId());
-                    deleteCvgOp = new DeleteCvgOperation(_(viewType.getNeighbors()).keys().value());
-                    _iwc.sendLocalNonOTOperation(CONFIG.WIDGET.NAME.VIEWCANVAS, deleteCvgOp.toNonOTOperation());
-                    var res= CVG(node,that.getRootSubjectEntity());
-                    var performCvgOp = new PerformCvgOperation(res);
-                    _iwc.sendLocalNonOTOperation(CONFIG.WIDGET.NAME.VIEWCANVAS, performCvgOp.toNonOTOperation());
-
+                    if(!fromCallback) {
+                        EntityManager.deleteFromMap(viewType.getViewId(), viewType.getEntityId());
+                        EntityManager.addToMap(viewType.getViewId(), node.getEntityId(), viewType.getEntityId());
+                        deleteCvgOp = new DeleteCvgOperation(_(viewType.getNeighbors()).keys().value());
+                        _iwc.sendLocalNonOTOperation(CONFIG.WIDGET.NAME.VIEWCANVAS, deleteCvgOp.toNonOTOperation());
+                        var res = CVG(node, that.getRootSubjectEntity());
+                        var performCvgOp = new PerformCvgOperation(res);
+                        _iwc.sendLocalNonOTOperation(CONFIG.WIDGET.NAME.VIEWCANVAS, performCvgOp.toNonOTOperation());
+                    }
 
                 }
             }
@@ -152,12 +164,12 @@ define([
 
         /**
          * Propagate a Value Change to the remote users and the local widgets
-		 * @param type Type of the update (CONFIG.OPERATION.TYPE.INSERT,DELETE)
-		 * @param value Char that was inserted or deleted
-		 * @param position Position the change took place
-		 */
-		this.propagateValueChange = function (type, value, position) {
-			var operation = new ValueChangeOperation(that.getEntityId(), value, type, position);
+         * @param type Type of the update (CONFIG.OPERATION.TYPE.INSERT,DELETE)
+         * @param value Char that was inserted or deleted
+         * @param position Position the change took place
+         */
+        this.propagateValueChange = function (type, value, position) {
+            var operation = new ValueChangeOperation(that.getEntityId(), value, type, position);
             if(that.getRootSubjectEntity().getViewId()) {
                 _iwc.setBufferedMessagesReceiver(CONFIG.WIDGET.NAME.VIEWCANVAS);
                 propagateValueChangeOperation(operation, CONFIG.WIDGET.NAME.VIEWCANVAS);
@@ -167,83 +179,84 @@ define([
                 propagateValueChangeOperation(operation, CONFIG.WIDGET.NAME.MAIN);
 
             }
-		};
-		/**
-		 * Propagate a Value Change Operation to the remote users and the local widgets
-		 * @param {operations.ot.ValueChangeOperation} operation
+        };
+
+        /**
+         * Propagate a Value Change Operation to the remote users and the local widgets
+         * @param {operations.ot.ValueChangeOperation} operation
          * @param {string} component the receiver
-		 */
-		var propagateValueChangeOperation = function (operation, component) {
-			processValueChangeOperation(operation);
+         */
+        var propagateValueChangeOperation = function (operation, component) {
+            processValueChangeOperation(operation);
             _iwc.sendLocalOTOperation(component, operation.getOTOperation());
-		};
+        };
 
-		/**
-		 * Callback for a Value Change Operation
-		 * @param {operations.ot.ValueChangeOperation} operation
-		 */
-		var valueChangeCallback = function (operation) {
-			if (operation instanceof ValueChangeOperation && operation.getEntityId() === that.getEntityId()) {
-				processValueChangeOperation(operation, true);
-			}
-		};
+        /**
+         * Callback for a Value Change Operation
+         * @param {operations.ot.ValueChangeOperation} operation
+         */
+        var valueChangeCallback = function (operation) {
+            if (operation instanceof ValueChangeOperation && operation.getEntityId() === that.getEntityId()) {
+                processValueChangeOperation(operation, true);
+            }
+        };
 
-		var init = function () {
-			_$node.off();
-			_$node.change(function () {
-				that.propagateValueChange(CONFIG.OPERATION.TYPE.UPDATE, $(this).val(), 0);
-			});
-		};
+        var init = function () {
+            _$node.off();
+            _$node.change(function () {
+                that.propagateValueChange(CONFIG.OPERATION.TYPE.UPDATE, $(this).val(), 0);
+            });
+        };
 
-		/**
-		 * Set value
-		 * @param {string} value
-		 */
-		this.setValue = function (value) {
-			_value = value;
-			_$node.val(value);
-		};
+        /**
+         * Set value
+         * @param {string} value
+         */
+        this.setValue = function (value) {
+            _value = value;
+            _$node.val(value);
+        };
 
-		/**
-		 * Get value
-		 * @returns {string}
-		 */
-		this.getValue = function () {
-			return _value;
-		};
+        /**
+         * Get value
+         * @returns {string}
+         */
+        this.getValue = function () {
+            return _value;
+        };
 
-		/**
-		 * Get jQuery object of DOM node representing the value
-		 * @returns {jQuery}
-		 */
-		this.get$node = function () {
-			return _$node;
-		};
+        /**
+         * Get jQuery object of DOM node representing the value
+         * @returns {jQuery}
+         */
+        this.get$node = function () {
+            return _$node;
+        };
 
-		/**
-		 * Register inter widget communication callbacks
-		 */
-		this.registerCallbacks = function () {
-			_iwc.registerOnDataReceivedCallback(valueChangeCallback);
-		};
+        /**
+         * Register inter widget communication callbacks
+         */
+        this.registerCallbacks = function () {
+            _iwc.registerOnDataReceivedCallback(valueChangeCallback);
+        };
 
-		/**
-		 * Unregister inter widget communication callbacks
-		 */
-		this.unregisterCallbacks = function () {
-			_iwc.unregisterOnDataReceivedCallback(valueChangeCallback);
-		};
+        /**
+         * Unregister inter widget communication callbacks
+         */
+        this.unregisterCallbacks = function () {
+            _iwc.unregisterOnDataReceivedCallback(valueChangeCallback);
+        };
 
-		this.setValueFromJSON = function (json) {
-			this.setValue(json.value);
-		};
-		init();
+        this.setValueFromJSON = function (json) {
+            this.setValue(json.value);
+        };
+        init();
 
-		if (_iwc) {
-			that.registerCallbacks();
-		}
-	}
+        if (_iwc) {
+            that.registerCallbacks();
+        }
+    }
 
-	return SelectionValue;
+    return SelectionValue;
 
 });
